+++ packages/Button/src/components/Link/LinkButton.styles.js
@@ -1,8 +1,15 @@
 import styled, { css } from "styled-components";
 import tokens from "@paprika/tokens";
 import stylers from "@paprika/stylers";
-import { commonStyles, sizeStyles, kindStyles, activeStyles, inactiveStyles, fullWidthStyles } from "./Button.styles";
-import { ButtonIcon } from "./Button";
+import {
+  commonStyles,
+  sizeStyles,
+  kindStyles,
+  activeStyles,
+  inactiveStyles,
+  fullWidthStyles,
+} from "../../Button.styles";
+import { ButtonIcon } from "../../Button";
 
 export const LinkButton = styled.a(
   ({ size, kind, isFullWidth, isActive, ...props }) => css`