+++ packages/Button/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+Move CloseButton, IconButton, LinkButton to components folder to fix typings and for file structure consistency [kcumlat](https://github.com/kcumlat).
+
 ## [0.3.0] - 2020-06-16
 
 ### Changed